+++ src/simd/i32x.rs
@@ -14,6 +14,13 @@ pub type I32x<const N: usize> = Simd<i32, N>;
 pub type I32x2 = Simd<i32, 2>;
 pub type I32x4 = Simd<i32, 4>;
 
+impl<const N: usize> Default for I32x<N> where (): SimdLanes<N> {
+    #[inline(always)]
+    fn default() -> Self {
+        Self::splat(Default::default())
+    }
+}
+
 impl<const N: usize> I32x<N> where (): SimdLanes<N> {
     #[allow(non_snake_case)]
     #[inline(always)]